Move some test functions from androidTest to sharedTest.
Some functions in these test files aren't yet ready to be run on robolectric. Only move the test functions that can run on robolectric for now.
