+++ app/src/sharedTest/java/ca/rmen/android/poetassistant/shared/main/ShareTest.java
@@ -17,7 +17,7 @@
  * along with Poet Assistant.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package ca.rmen.android.poetassistant.main;
+package ca.rmen.android.poetassistant.shared.main;
 
 
 import static androidx.test.espresso.Espresso.onView;
@@ -25,22 +25,18 @@ import static androidx.test.espresso.action.ViewActions.click;
 import static androidx.test.espresso.intent.Intents.intended;
 import static androidx.test.espresso.intent.matcher.IntentMatchers.hasAction;
 import static androidx.test.espresso.intent.matcher.IntentMatchers.hasExtra;
-import static androidx.test.espresso.matcher.ViewMatchers.hasSibling;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
-import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
 import static org.hamcrest.Matchers.allOf;
 import static org.hamcrest.Matchers.containsString;
+import static org.hamcrest.Matchers.endsWith;
 import static org.hamcrest.Matchers.is;
-import static ca.rmen.android.poetassistant.main.TestAppUtils.addFilter;
 import static ca.rmen.android.poetassistant.main.TestAppUtils.search;
-import static ca.rmen.android.poetassistant.main.TestAppUtils.typeAndSpeakPoem;
 import static ca.rmen.android.poetassistant.main.TestUiUtils.openMenuItem;
 import static ca.rmen.android.poetassistant.main.TestUiUtils.swipeViewPagerLeft;
 
 import android.content.Context;
 import android.content.Intent;
-import android.os.Build;
 
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.filters.LargeTest;
@@ -51,6 +47,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import ca.rmen.android.poetassistant.R;
+import ca.rmen.android.poetassistant.main.MainActivity;
 import ca.rmen.android.poetassistant.main.rules.PoetAssistantIntentsTestRule;
 import ca.rmen.android.poetassistant.main.rules.RetryTestRule;
 
@@ -64,55 +61,45 @@ public class ShareTest {
     @Rule
     public PoetAssistantIntentsTestRule<MainActivity> mActivityTestRule = new PoetAssistantIntentsTestRule<>(MainActivity.class);
 
-    @Test
-    public void sharePoemTest() {
-        swipeViewPagerLeft(3);
-        String poemText = "Let us share a poem";
-        typeAndSpeakPoem(poemText);
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
-            openMenuItem(R.string.share_poem_text);
-        } else {
-            openMenuItem(R.string.share);
-        }
-        checkShareIntentEquals(poemText);
-    }
 
     @Test
-    public void shareThesaurusTest() {
-        search("splurge");
-        swipeViewPagerLeft(1);
+    public void shareRhymesTest() {
+        search("merge");
         openMenuItem(R.string.share);
-        checkShareIntentContains("flaunt");
+        checkShareIntentContains("upsurge");
+    }
+
+
+    @Test
+    public void shareDictionaryTest() {
+        search("a");
+        swipeViewPagerLeft(2);
+        openMenuItem(R.string.share);
+        checkShareIntentContains("the blood group whose red cells carry the A antigen");
     }
 
     @Test
-    public void shareFilteredThesaurusTest() {
+    public void sharePatternTest() {
+        search("ho?t");
+        openMenuItem(R.string.share);
+        checkShareIntentContains("host");
+    }
+
+    @Test
+    public void shareWotdTest() {
         Context context = mActivityTestRule.getActivity();
-        search("happy");
-        swipeViewPagerLeft(1);
-        addFilter("messed", "blessed");
+        openMenuItem(R.string.action_wotd_history);
         openMenuItem(R.string.share);
-        String expectedContent = context.getString(R.string.share_thesaurus_title_with_filter, "happy", "messed");
-        checkShareIntentContains(expectedContent);
+        checkShareIntentContains(context.getString(R.string.share_wotd_title));
     }
 
     @Test
-    public void shareFavoritesTest() {
-        search("happy");
-        onView(allOf(withId(R.id.btn_star_result), isDisplayed(), hasSibling(withText("snappy")))).perform(click());
-        onView(allOf(withId(R.id.btn_star_result), isDisplayed(), hasSibling(withText("crappy")))).perform(click());
-        swipeViewPagerLeft(4);
-        openMenuItem(R.string.share);
-        checkShareIntentContains("snappy");
-    }
-
-    // Need to look at this: sometimes the app bar layout is hidden :(
-    @Ignore
-    @Test
-    public void shareRandomWordTest() {
-        openMenuItem(R.string.action_random_word);
-        openMenuItem(R.string.share);
-        checkShareIntentContains("Definitions of");
+    public void sharePopupTest() {
+        search("strawberry");
+        Context context = mActivityTestRule.getActivity();
+        onView(allOf(withText("adversary"), isDisplayed())).perform(click());
+        onView(allOf(withText(endsWith(context.getString(R.string.share))), isDisplayed())).perform(click());
+        checkShareIntentEquals("adversary");
     }
 
     private void checkShareIntentContains(String expectedText) {
